+++ src/installer/tests/Microsoft.NET.HostModel.Tests/AppHost.Bundle.Tests/SingleFileApiTests.cs
@@ -8,11 +8,12 @@ using Xunit;
 
 namespace AppHost.Bundle.Tests
 {
-    public class SingleFileApiTests : BundleTestBase, IClassFixture<SingleFileApiTests.SharedTestState>
+    [Collection(nameof(SingleFileSharedCollection))]
+    public class SingleFileApiTests : BundleTestBase
     {
-        private SharedTestState sharedTestState;
+        private SingleFileSharedState sharedTestState;
 
-        public SingleFileApiTests(SharedTestState fixture)
+        public SingleFileApiTests(SingleFileSharedState fixture)
         {
             sharedTestState = fixture;
         }
@@ -119,24 +120,6 @@ namespace AppHost.Bundle.Tests
                 .Should().Pass()
                 .And.HaveStdOutContaining(extractionDir)
                 .And.HaveStdOutContaining(bundleDir);
-        }
-
-        public class SharedTestState : SharedTestStateBase, IDisposable
-        {
-            public TestProjectFixture TestFixture { get; set; }
-
-            public SharedTestState()
-            {
-                // We include mockcoreclr in our project to test native binaries extraction.
-                string mockCoreClrPath = Path.Combine(RepoDirectories.Artifacts, "corehost_test",
-                    RuntimeInformationExtensions.GetSharedLibraryFileNameForCurrentPlatform("mockcoreclr"));
-                TestFixture = PreparePublishedSelfContainedTestProject("SingleFileApiTests", $"/p:AddFile={mockCoreClrPath}");
-            }
-
-            public void Dispose()
-            {
-                TestFixture.Dispose();
-            }
         }
     }
 }
